Fix -Wmissing-prototypes

+++ firmware/tlc5940-lpc812/utils.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-uint32_t random_min_max(uint16_t min, uint16_t max);
+uint16_t random_min_max(uint16_t min, uint16_t max);
 
 #endif // __UTILS_H
